requirements.txt
python-twitter
requests
TextBlob
nltk
